Flatten this struct into the enum
- This is an experiment to see how much i like it. This moves the
declaration of Comparison closer together with the NumberComparison
and hopefully reduces the complexity with keeping all these structs
connected in your brain.
We cant do this with the NumberComparator since this is another
enum. Perhaps there is a way to refactor this as well since the map uses
a similar construct.